+++ tests/testthat/test-db.R
@@ -6,7 +6,6 @@ context("db")
 
 test_that("neon_db", {
   
-  
   db <- neon_db()
   expect_is(db, "DBIConnection")
 
@@ -14,11 +13,11 @@ test_that("neon_db", {
   
   ## Confirm cached connection
   expect_identical(db, db2)
-  
   neon_disconnect(db)
+  gc()
+
   
   neon_delete_db(ask = FALSE)
-  
 })
 
 
@@ -126,10 +125,10 @@ test_that("neon_table", {
 
 test_that("check_tablename", {
   
-  expect_error(check_tablename("A", c("B", "D")))
-  expect_error(check_tablename("A", c("A-basic", "A-expanded")))
+  expect_error(check_tablename("A", tables = c("B", "D")))
+  expect_error(check_tablename("A", tables =  c("A-basic", "A-expanded")))
   
-  x <- check_tablename("A-expan", c("A-basic", "A-expanded"))
+  x <- check_tablename("A-expan",  tables = c("A-basic", "A-expanded"))
   expect_equal(x, "A-expanded")
   
 })